fix: attr not found error for H5Aiterate
dataseti_open() is called inside H5Aiterate() but dataseti_open does
not check whether this file is regular or log-based.

Originally, file type checking (regular vs log-based) is implemented
in dataset_open(). In this commit, the checking is moved from
dataset_open to dataseti_open so that both 1) dataset_open and 2)
H5Aiterate can make use of it.