H5VL_log_nb_flush_write_reqs does not need dxplid argument
1. H5VL_log_nb_flush_write_reqs will create a new dxplid and set
collective MPI-IO anyway. There's no need to pass in a dxplid as an
argument.
2. H5VL_log_nb_flush_write_reqs is called either in H5Fflush or
H5Fclose. In either case, user will not be able to pass a cutomized
dxplid. So the original dxplid argument is not necessary since
it is always a default dxplid (or a copy of a defualt dxplid).