fix: subfile read: mismatch of nsubfiles and nproc
It is possible that, for example, a file is created with 8 subfiles.
But when openning and reading the file, we only use 4 processes.

In the original implementation before this fix, the info of 8 subfiles
is not saved. Only the first fp->ngroup subfiles will be opened for
read, where fp->ngroup is a number bounded by the number of processes
(i.e. <= 4 in this case).

In this fix, we use fp->nsubfiles to store the number of subfiles for an
opened file. All fp->nsubfiles subfiles will be opened for read.