testcase: subfile_dread tests multiple scenarios
Test the following:
1. nsbufile > nproc
2. nsubfile == nproc
3. nsubfile < nproc

For each of the above, test:
1. read pattern same as write pattern (row wise)
2. read pattern is row wise, but each process read a different row than
it writes. (read from one subfile that process is not responsible for)
3. read pattern is column wise (read from several subfiles)
4. read all dataset.

This means a total of 12 scenarios are tested.

Also, we test each scenario using 1 to 12 number of processes. This
makes sures Log VOL also works for odd number of processes.